Code clean-up 2

+++ src/guilayer/employees/ListEmployees.java
@@ -1,24 +1,17 @@
 package guilayer.employees;
 
-import java.util.ArrayList;
-import java.util.function.Consumer;
-
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
-import javax.swing.table.AbstractTableModel;
 
 import ctrllayer.EmployeeController;
 import guilayer.ManagerWindow;
 import guilayer.interfaces.ButtonColumn;
+import guilayer.interfaces.ItemTableModel;
 import guilayer.interfaces.PerformListener;
-import guilayer.inventory.ListInventory;
 import modlayer.Employee;
-import modlayer.Unit;
 
 import javax.swing.ListSelectionModel;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
 import javax.swing.JButton;
 import javax.swing.JOptionPane;
 
@@ -28,13 +21,7 @@ import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.awt.event.ActionEvent;
 import javax.swing.JTextField;
-import javax.swing.JComboBox;
-import javax.swing.DefaultComboBoxModel;
-import javax.swing.JRadioButton;
-import javax.swing.JList;
 import javax.swing.AbstractAction;
-import javax.swing.AbstractListModel;
-import javax.swing.Action;
 import javax.swing.event.CaretListener;
 import javax.swing.event.CaretEvent;
 
@@ -85,15 +72,18 @@ public class ListEmployees extends JPanel implements ActionListener, MouseListen
 		table.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
 		table.getTableHeader().setReorderingAllowed(false);
 		table.setAutoCreateRowSorter(true);
-		model.setEmployees(employeeCtrl.getEmployees());
+		model.setItems(employeeCtrl.getEmployees());
 		table.setModel(model);
 		scrollPane.setViewportView(table);
 
 		AbstractAction delete = new AbstractAction() {
 			public void actionPerformed(ActionEvent e) {
-				JTable table = (JTable) e.getSource();
+				if (JOptionPane.showConfirmDialog(ListEmployees.this, "Are you sure?") != JOptionPane.YES_OPTION) {
+		        	return;
+		        }
+				
 				int modelRowIndex = Integer.valueOf(e.getActionCommand());
-				Employee employee = model.getEmployeeAt(modelRowIndex);
+				Employee employee = model.getItem(modelRowIndex);
 
 				if (!employeeCtrl.deleteEmployee(employee)) {
 					JOptionPane.showMessageDialog(ListEmployees.this, "An error occured while deleting the Employee!",
@@ -103,7 +93,7 @@ public class ListEmployees extends JPanel implements ActionListener, MouseListen
 
 				JOptionPane.showMessageDialog(ListEmployees.this, "The Employee was successfully deleted!", "Success!",
 						JOptionPane.INFORMATION_MESSAGE);
-				model.setEmployees(employeeCtrl.getEmployees());
+				reset();
 			}
 		};
 
@@ -115,97 +105,71 @@ public class ListEmployees extends JPanel implements ActionListener, MouseListen
 		btn_create.addActionListener(this);
 		table.addMouseListener(this);
 	}
-
-	private void searchInventory() {
-
-		String keyword = txt_search.getText().trim();
-		model.setEmployees(employeeCtrl.searchEmployees(keyword));
+	private void reset() {
+		model.setItems(employeeCtrl.getEmployees());
+		txt_search.setText("");
 	}
-
-	@Override
-	public void caretUpdate(CaretEvent e) {
-		if (e.getSource() == txt_search) {
-			searchInventory();
-		}
+	private void search() {
+		String keyword = txt_search.getText().trim();
+		model.setItems(employeeCtrl.searchEmployees(keyword));
 	}
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		if (e.getSource() == btn_search) {
-			searchInventory();
+			search();
 		}
 		if (e.getSource() == btn_create) {
 			setVisible(false);
-			employeeEditor.createEmployee();
+			employeeEditor.create();
 		}
 	}
-
 	@Override
 	public void mouseClicked(MouseEvent e) {
 		if (e.getClickCount() == 2 && e.getSource() == table) {
 			int viewRowIndex = table.getSelectedRow();
 			int modelRowIndex = table.convertRowIndexToModel(viewRowIndex);
-			Employee employee = model.getEmployeeAt(modelRowIndex);
+			Employee employee = model.getItem(modelRowIndex);
 
+			employeeEditor.update(employee);
 			setVisible(false);
-			employeeEditor.updateEmployee(employee);
+		}
+	}
+	@Override
+	public void caretUpdate(CaretEvent e) {
+		if (e.getSource() == txt_search) {
+			search();
 		}
 	}
 	@Override
 	public void performed() {
-		model.setEmployees(employeeCtrl.getEmployees());
+		model.setItems(employeeCtrl.getEmployees());
 		setVisible(true);
 	}
-
 	@Override
 	public void cancelled() {
 		setVisible(true);
 	}
-
 	@Override
-	public void mousePressed(MouseEvent e) {
-	}
-
+	public void mousePressed(MouseEvent e) {}
 	@Override
-	public void mouseReleased(MouseEvent e) {
-	}
-
+	public void mouseReleased(MouseEvent e) {}
 	@Override
-	public void mouseEntered(MouseEvent e) {
-	}
-
+	public void mouseEntered(MouseEvent e) {}
 	@Override
-	public void mouseExited(MouseEvent e) {
-	}
+	public void mouseExited(MouseEvent e) {}
 
-	private class EmployeeTableModel extends AbstractTableModel {
-
-		private String[] columns = new String[] { "CPR", "Name", "Username", "Address", "Zip code", "City", "Phone", "Email", "Access", "" };
-		private ArrayList<Employee> employees;
+	private class EmployeeTableModel extends ItemTableModel<Employee> {
 
 		public EmployeeTableModel() {
-			this(new ArrayList<Employee>());
+			super();
+			
+			columns = new String[] { "CPR", "Name", "Username", "Address", "Zip code", "City", "Phone", "Email", "Access", "" };
 		}
-
-		public EmployeeTableModel(ArrayList<Employee> employees) {
-			this.employees = employees;
-			update();
-		}
-
-		@Override
-		public int getRowCount() {
-			return employees.size();
-		}
-
-		@Override
-		public int getColumnCount() {
-			return columns.length;
-		}
-
+		
 		@Override
 		public Object getValueAt(int rowIndex, int columnIndex) {
-
-			Employee employee = employees.get(rowIndex);
+			Employee employee = items.get(rowIndex);
 
 			switch (columnIndex) {
 			case 0:
@@ -232,56 +196,9 @@ public class ListEmployees extends JPanel implements ActionListener, MouseListen
 
 			return null;
 		}
-
-		@Override
-		public String getColumnName(int columnIndex) {
-			return columns[columnIndex];
-		}
-
-		@Override
-		public Class getColumnClass(int columnIndex) {
-			switch (columnIndex) {
-			case 0:
-				return String.class;
-			case 1:
-				return String.class;
-			case 2:
-				return String.class;
-			case 3:
-				return String.class;
-			case 4:
-				return String.class;
-			case 5:
-				return String.class;
-			case 6:
-				return String.class;
-			case 7:
-				return String.class;
-			case 8:
-				return Integer.class;
-			case 9:
-				return JButton.class;
-			}
-
-			return null;
-		}
-
 		@Override
 		public boolean isCellEditable(int rowIndex, int columnIndex) {
 			return columnIndex == getColumnCount() - 1;
-		}
-
-		public void update() {
-			fireTableDataChanged();
-		}
-
-		public Employee getEmployeeAt(int rowIndex) {
-			return employees.get(rowIndex);
-		}
-
-		public void setEmployees(ArrayList<Employee> employees) {
-			this.employees = employees;
-			update();
 		}
 	}
 }
